electric-service-automation-main/backend/orders-api/requirements.txt
pydantic==2.5.0
psycopg2-binary==2.9.9
